backend/pyproject.toml
[project]
name = "evolution-of-todo-backend"
version = "0.1.0"
description = "Phase II Backend - Python REST API for Todo Application"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "sqlmodel>=0.0.22",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "bcrypt>=4.1.2",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.3",
    "httpx>=0.25.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/"]

[tool.hatch.build.targets.sdist]
include = ["/src", "/tests"]
